+++ src/app/app.module.ts
@@ -62,7 +62,7 @@ import { QuillModule } from 'ngx-quill';
     BrowserAnimationsModule,
     HttpClientModule,
     SharedModule,
-    QuillModule.forRoot()
+    QuillModule.forRoot(),
   ],
   providers: [
     {