+++ src/app/modules/project/project.module.ts
@@ -28,15 +28,17 @@ import { EmbedButtonComponent } from './embed-button/embed-button.component';
 import { EmbedComponent } from './embed/embed.component';
 import { PaginationModule } from 'ngx-bootstrap/pagination';
 import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
-import { ModalHighlightDeleteComponent } from 'src/app/modules/project/modal-highlight-delete/modal-highlight-delete.component';
-import { ModalHighlightComponent } from 'src/app/modules/project/modal-highlight/modal-highlight.component';
-import { SharedModule } from 'src/app/modules/shared/shared.module';
+import { ModalHighlightFormComponent } from 'src/app/modules/project/modal-highlight-form/modal-highlight-form.component';
+import { SharedModule } from './../shared/shared.module';
 import { QuillModule } from 'ngx-quill';
 import { SafeHtmlPipe } from 'src/app/utils/safeHtml.pipe';
 import { FileUploaderComponent } from 'src/app/components/file-uploader/file-uploader.component';
 import { DndDirective } from 'src/app/components/file-uploader/DndDirective';
 import { Meta } from '@angular/platform-browser';
+import { FormatDatePipe } from 'src/app/utils/format-date.pipe';
+import { HighlightsModalComponent } from './highlights-modal/highlights-modal.component';
 import { ProjectComponent } from 'src/app/modules/project/overview/project/project.component';
+import { NgxDebounceClickModule } from '@ngx-lite/debounce-click';
 import { PopoverModule } from 'ngx-bootstrap/popover';
 import { CollaboratorComponent } from './collaborator/collaborator.component';
 
@@ -47,11 +49,13 @@ import { CollaboratorComponent } from './collaborator/collaborator.component';
     EditComponent,
     EmbedButtonComponent,
     EmbedComponent,
-    ModalHighlightDeleteComponent,
-    ModalHighlightComponent,
-    FileUploaderComponent,
+    ModalHighlightFormComponent,
+    FormatDatePipe,
     DndDirective,
+    FileUploaderComponent,
     SafeHtmlPipe,
+    HighlightsModalComponent,
+    ProjectComponent,
     CollaboratorComponent,
     ProjectComponent
   ],
@@ -66,6 +70,8 @@ import { CollaboratorComponent } from './collaborator/collaborator.component';
     BsDropdownModule.forRoot(),
     ModalModule.forRoot(),
     TooltipModule.forRoot(),
+    QuillModule,
+    NgxDebounceClickModule,
     PopoverModule.forRoot(),
     QuillModule
   ],