+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Added
-
+- Added update highlight functionality - [#294](https://github.com/DigitalExcellence/dex-frontend/issues/294)
 
 ### Changed